Improving lua
