--- coc-settings.json
@@ -1,15 +1,0 @@
-{
-    "clangd.arguments": [
-        "--compile-commands-dir=./build",
-        "--background-index",
-        "--completion-style=detailed"
-    ],
-    "clangd.fallbackFlags": [
-        "-Wall"
-    ],
-    "coc.preferences.semanticTokensHighlights": false,
-    "clangd.semanticHighlighting": true,
-    "rome.enable": true,
-    "jedi.enable": true,
-    "jedi.trace.server": "verbose"
-}
