+++ lua/g.lua
@@ -1,0 +1,2 @@
+-- Prevents default windows
+vim.g.windowswap_map_keys = 0
